Add missing snapshot dependencies to the generated docker images (#33)
Spring Boot's extract tool that the Dockerfiles use to extract files from the spring boot jar file handles snapshot and non-snapshot dependencies differently - which currently causes the snapshot dependencies to be missing from the created docker image. With this commit, the snapshot dependencies are also added to the created docker image.

Resolves #30